Refactor code.

+++ src/main/java/edu/xaut/CalculatorStrategy.java
@@ -2,5 +2,5 @@ package edu.xaut;
 
 public interface CalculatorStrategy {
 
-    int calculate(int distance, long time);
+    float calculate(int distance, long time);
 }
